main/requirements.txt
langchain==0.2.9
langchain_community==0.2.7
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.36.0
faiss-cpu
sentence-transformers